third_party/WebKit/LayoutTests/external/wpt/streams/writable-streams/constructor.dedicatedworker-expected.txt

This is a testharness.js-based test.
PASS controller argument should be passed to start method
PASS controller argument should be passed to write method
PASS controller argument should not be passed to close method
PASS highWaterMark should be reflected to desiredSize
PASS WritableStream should be writable and ready should fulfill immediately if the strategy does not apply backpressure
PASS WritableStream should be constructible with no arguments
PASS WritableStream instances should have standard methods and properties
PASS private constructors should not be exported
PASS WritableStreamDefaultController constructor should throw
PASS WritableStreamDefaultController constructor should throw when passed an initialised WritableStream
PASS WritableStreamDefaultWriter should throw unless passed a WritableStream
PASS WritableStreamDefaultWriter constructor should throw when stream argument is locked
FAIL WritableStream constructor should stop after get on size fails assert_equals: operations should be performed in the right order expected "get on size" but got "get on type,get on size"
FAIL WritableStream constructor should stop after get on highWaterMark fails assert_equals: operations should be performed in the right order expected "get on size,get on highWaterMark" but got "get on type,get on size,get on highWaterMark"
FAIL WritableStream constructor should stop after get on type fails assert_equals: operations should be performed in the right order expected "get on size,get on highWaterMark,get on type" but got "get on type"
FAIL WritableStream constructor should stop after validate on type fails assert_equals: operations should be performed in the right order expected "get on size,get on highWaterMark,get on type" but got "get on type,get on size,get on highWaterMark"
FAIL WritableStream constructor should stop after validate on size fails assert_equals: operations should be performed in the right order expected "get on size,get on highWaterMark,get on type" but got "get on type,get on size,get on highWaterMark"
FAIL WritableStream constructor should stop after tonumber on highWaterMark fails assert_equals: operations should be performed in the right order expected "get on size,get on highWaterMark,get on type,tonumber on highWaterMark" but got "get on type,get on size,get on highWaterMark,tonumber on highWaterMark"
FAIL WritableStream constructor should stop after validate on highWaterMark fails assert_equals: operations should be performed in the right order expected "get on size,get on highWaterMark,get on type,tonumber on highWaterMark" but got "get on type,get on size,get on highWaterMark,tonumber on highWaterMark"
FAIL WritableStream constructor should stop after get on write fails assert_equals: operations should be performed in the right order expected "get on size,get on highWaterMark,get on type,tonumber on highWaterMark,get on write" but got "get on type,get on size,get on highWaterMark,tonumber on highWaterMark,get on write"
FAIL WritableStream constructor should stop after validate on write fails assert_equals: operations should be performed in the right order expected "get on size,get on highWaterMark,get on type,tonumber on highWaterMark,get on write" but got "get on type,get on size,get on highWaterMark,tonumber on highWaterMark,get on write"
FAIL WritableStream constructor should stop after get on close fails assert_equals: operations should be performed in the right order expected "get on size,get on highWaterMark,get on type,tonumber on highWaterMark,get on write,get on close" but got "get on type,get on size,get on highWaterMark,tonumber on highWaterMark,get on write,get on close"
FAIL WritableStream constructor should stop after validate on close fails assert_equals: operations should be performed in the right order expected "get on size,get on highWaterMark,get on type,tonumber on highWaterMark,get on write,get on close" but got "get on type,get on size,get on highWaterMark,tonumber on highWaterMark,get on write,get on close"
FAIL WritableStream constructor should stop after get on abort fails assert_equals: operations should be performed in the right order expected "get on size,get on highWaterMark,get on type,tonumber on highWaterMark,get on write,get on close,get on abort" but got "get on type,get on size,get on highWaterMark,tonumber on highWaterMark,get on write,get on close,get on abort"
FAIL WritableStream constructor should stop after validate on abort fails assert_equals: operations should be performed in the right order expected "get on size,get on highWaterMark,get on type,tonumber on highWaterMark,get on write,get on close,get on abort" but got "get on type,get on size,get on highWaterMark,tonumber on highWaterMark,get on write,get on close,get on abort"
FAIL WritableStream constructor should stop after get on start fails assert_equals: operations should be performed in the right order expected "get on size,get on highWaterMark,get on type,tonumber on highWaterMark,get on write,get on close,get on abort,get on start" but got "get on type,get on size,get on highWaterMark,tonumber on highWaterMark,get on write,get on close,get on abort,get on start"
FAIL WritableStream constructor should stop after validate on start fails assert_equals: operations should be performed in the right order expected "get on size,get on highWaterMark,get on type,tonumber on highWaterMark,get on write,get on close,get on abort,get on start" but got "get on type,get on size,get on highWaterMark,tonumber on highWaterMark,get on write,get on close,get on abort,get on start"
Harness: the test ran to completion.
